change grub to not prompt you after a dirty shutdown.
If there is a power outage during the boot process the next boot will be
halted at grub menu. requiring a keyboard to be hooked up to boot. lets
stop that as this is just a media center pc with non-esentail data on it,
foring a boot after a dirty shutdown/boot should be okay in this situation