Add new function "InstallLTSEnablementStack".
This new function will check if your OS is equal to 12.04, If so it will
prompt the user asking if they want to install the LTSEnablementStack. I
suggest this as it installs the 3.8 kernel which fixes a lot of bugs (at
least for me) On any other install besides from the minimal iso would
include this by default. Still defaults to "No" though just for compatibility