XbmcUpstartScript: General update (read more)...
When playing around with a new PC I was having issues with XBMC not starting
every boot. it would just sit at the tty prompt and X would be a defunct
process. After some digging I came to the conclusion there was a race
condition going on with some of the hardware. A log o the boot process
revealed XBMC trying to start pretty early in the boot process Changing
the start on rule to on runlevel 2345 fixed the issue instantly, and a log
of the boot shows it booting way further down. This is the standard way
to start a service that reguires the whole system to be up i found ount
so this is actually the correct way.

Also disabled the xcuroor as XBMCs still shows up and it avoids having
that ugly black "X" pop up in the middle o the XBMC splash at boot :)